utils: Move iterator helpers to their own module
In the interest of cleanliness.
